refactor: 예약 step1 activity fragment로 변경
