+++ app/src/main/AndroidManifest.xml
@@ -64,9 +64,6 @@
             android:name=".ui.reservation.ReservationStep2Activity"
             android:exported="false" />
         <activity
-            android:name=".ui.reservation.ReservationStep1Activity"
-            android:exported="false" />
-        <activity
             android:name=".ui.main.companion.LiveCompanionActivity"
             android:exported="false" />
         <activity
