+++ app/src/main/java/com/example/team25/ui/reservation/ReservationActivity.kt
@@ -8,5 +8,11 @@ class ReservationActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_reservation)
+
+        if (savedInstanceState == null) {
+            supportFragmentManager.beginTransaction()
+                .replace(R.id.fragment_container_view, ReservationStep1Fragment())
+                .commit()
+        }
     }
 }
