+++ app/src/main/res/layout/activity_reservation.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.reservation.ReservationActivity">
+    tools:context=".ui.reservation.ReservationStep1Fragment">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment_container_view"
